+++ CHANGELOG.md
@@ -10,5 +10,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Presentation layer with Web API.
-- Basic structure of a Web API
-- Dockerfile for easy deployment at a later date
+- Basic structure of a Web API.
+- Dockerfile for easy deployment at a later date.
+
+## [1.1.0] - 2025-01-27
+
+### Added
+
+- Infrastructure and Presentation layers.
+- Service extension for dependency injection.
+- Database context.
+- Context implemented at project level.
+- Material, MaterialsLocation and World entities.
+- Database script “db.sql”.
+- Nuget packages installed in Infrastructure layer: Microsoft.EntityFrameworkCore, Microsoft.EntityFrameworkCore.Tools y Npgsql.EntityFrameworkCore.PostgreSQL.